+++ NEWS.md
@@ -17,6 +17,8 @@
 * Fix `parse_gcov` bug when package is stored in directory with regex special characters, see #459
 * Error/warning thrown for, respectively, missing gcov or empty parsed gcov output (@stephematician, #448)
 
+* Support Google Cloud Build uploading reports to Codecov.io (@MarkEdmondson1234 #469)
+
 * covr is now licensed as MIT (#454)
 
 # covr 3.5.1